+++ src/internal/SincResampler.h
@@ -16,21 +16,21 @@ public:
     // scaleFactor == sourceSampleRate / destinationSampleRate
     // kernelSize can be adjusted for quality (higher is better)
     // numberOfKernelOffsets is used for interpolation and is the number of sub-sample kernel shifts.
-    SincResampler(double scaleFactor, size_t kernelSize = 32, size_t numberOfKernelOffsets = 32);
+    SincResampler(double scaleFactor, int kernelSize = 32, int numberOfKernelOffsets = 32);
 
     // Processes numberOfSourceFrames from source to produce numberOfSourceFrames / scaleFactor frames in destination.
-    void process(const float * source, float * destination, size_t numberOfSourceFrames);
+    void process(const float * source, float * destination, int numberOfSourceFrames);
 
     // Process with input source callback function for streaming applications.
-    void process(AudioSourceProvider *, float * destination, size_t framesToProcess);
+    void process(AudioSourceProvider *, float * destination, int framesToProcess);
 
 protected:
     void initializeKernel();
-    void consumeSource(float * buffer, size_t numberOfSourceFrames);
+    void consumeSource(float * buffer, int numberOfSourceFrames);
 
     double m_scaleFactor;
-    size_t m_kernelSize;
-    size_t m_numberOfKernelOffsets;
+    int m_kernelSize;
+    int m_numberOfKernelOffsets;
 
     // m_kernelStorage has m_numberOfKernelOffsets kernels back-to-back, each of size m_kernelSize.
     // The kernel offsets are sub-sample shifts of a windowed sinc() shifted from 0.0 to 1.0 sample.
@@ -41,13 +41,13 @@ protected:
     double m_virtualSourceIndex{0};
 
     // This is the number of destination frames we generate per processing pass on the buffer.
-    size_t m_blockSize{512};
+    int m_blockSize{512};
 
     // Source is copied into this buffer for each processing pass.
     AudioFloatArray m_inputBuffer;
 
     const float * m_source{nullptr};
-    size_t m_sourceFramesAvailable{0};
+    int m_sourceFramesAvailable{0};
 
     // m_sourceProvider is used to provide the audio input stream to the resampler.
     AudioSourceProvider * m_sourceProvider{nullptr};
