+++ src/internal/src/AudioUtilities.cpp
@@ -25,8 +25,8 @@ double lab::AudioUtilities::discreteTimeConstantForSampleRate(double timeConstan
     return 1.0 - std::exp(-1.0 / (sampleRate * timeConstant));
 }
 
-size_t lab::AudioUtilities::timeToSampleFrame(double time, double sampleRate)
+int lab::AudioUtilities::timeToSampleFrame(double time, double sampleRate)
 {
     ASSERT(time >= 0);
-    return static_cast<size_t>(std::round(time * sampleRate));
+    return static_cast<int>(std::round(time * sampleRate));
 }
